+++ app/core/Engine/controllers/AccountsController/constants.ts
@@ -1,7 +1,4 @@
 import {
-  AccountsController,
-  AccountsControllerMessenger,
-  AccountsControllerState,
   AccountsControllerGetAccountByAddressAction as AccountsControllerGetAccountByAddressActionType,
   AccountsControllerSetAccountNameAction as AccountsControllerSetAccountNameActionType,
   AccountsControllerSetSelectedAccountAction as AccountsControllerSetSelectedAccountActionType,
@@ -14,61 +11,6 @@ import {
   AccountsControllerAccountAddedEvent as AccountsControllerAccountAddedEventType,
   AccountsControllerAccountRenamedEvent as AccountsControllerAccountRenamedEventType,
 } from '@metamask/accounts-controller';
-import { ControllerMessenger } from '../Engine.types';
-import Logger from '../../../util/Logger';
-
-// Default AccountsControllerState
-export const defaultAccountsControllerState: AccountsControllerState = {
-  internalAccounts: {
-    accounts: {},
-    selectedAccount: '',
-  },
-};
-
-/**
- * Creates instance of AccountsController
- *
- * @param options.messenger - Controller messenger instance
- * @param options.initialState - Initial state of AccountsController
- * @returns - AccountsController instance
- */
-export const createAccountsController = ({
-  messenger,
-  initialState,
-}: {
-  messenger: ControllerMessenger;
-  initialState?: AccountsControllerState;
-}): AccountsController => {
-  let accountsController = {} as AccountsController;
-
-  try {
-    const accountsControllerMessenger: AccountsControllerMessenger =
-      messenger.getRestricted({
-        name: 'AccountsController',
-        allowedEvents: [
-          'SnapController:stateChange',
-          'KeyringController:accountRemoved',
-          'KeyringController:stateChange',
-        ],
-        allowedActions: [
-          'KeyringController:getAccounts',
-          'KeyringController:getKeyringsByType',
-          'KeyringController:getKeyringForAccount',
-        ],
-      });
-
-    accountsController = new AccountsController({
-      messenger: accountsControllerMessenger,
-      state: initialState ?? defaultAccountsControllerState,
-    });
-  } catch (error) {
-    // Report error while initializing AccountsController
-    // TODO: Direct to vault recovery to reset controller states
-    Logger.error(error as Error, 'Failed to initialize AccountsController');
-  }
-
-  return accountsController;
-};
 
 // Action types of AccountsController
 export const AccountsControllerGetAccountByAddressAction: AccountsControllerGetAccountByAddressActionType['type'] =
